+++ src/main/java/net/minecraft/block/BlockPotato.java
@@ -38,7 +38,7 @@ public class BlockPotato extends BlockCrops
 
     public AxisAlignedBB getBoundingBox(IBlockState state, IBlockAccess source, BlockPos pos)
     {
-        return POTATO_AABB[((Integer)state.getValue(this.getAgeProperty())).intValue()];
+        return POTATO_AABB[(Integer) state.getValue(this.getAgeProperty())];
     }
 
     @Override
