+++ src/main/java/net/minecraft/client/util/SuffixArray.java
@@ -133,7 +133,7 @@ public class SuffixArray<T>
     {
         for (int i2 = 0; i2 < this.suffixToT.size(); ++i2)
         {
-            LOGGER.debug("{} {}", Integer.valueOf(i2), this.getString(i2));
+            LOGGER.debug("{} {}", i2, this.getString(i2));
         }
 
         LOGGER.debug("");
@@ -152,7 +152,7 @@ public class SuffixArray<T>
                 stringbuilder.append('^');
             }
 
-            int l2 = ((Integer)this.chars.get(j2 + k2)).intValue();
+            int l2 = (Integer) this.chars.get(j2 + k2);
 
             if (l2 == -1)
             {
@@ -209,7 +209,7 @@ public class SuffixArray<T>
 
             if (DEBUG_PRINT_COMPARISONS)
             {
-                LOGGER.debug("comparing lower \"{}\" with {} \"{}\": {}", p_194055_1_, Integer.valueOf(l2), this.getString(l2), Integer.valueOf(i3));
+                LOGGER.debug("comparing lower \"{}\" with {} \"{}\": {}", p_194055_1_, l2, this.getString(l2), i3);
             }
 
             if (i3 > 0)
@@ -234,7 +234,7 @@ public class SuffixArray<T>
 
                 if (DEBUG_PRINT_COMPARISONS)
                 {
-                    LOGGER.debug("comparing upper \"{}\" with {} \"{}\": {}", p_194055_1_, Integer.valueOf(j4), this.getString(j4), Integer.valueOf(j3));
+                    LOGGER.debug("comparing upper \"{}\" with {} \"{}\": {}", p_194055_1_, j4, this.getString(j4), j3);
                 }
 
                 if (j3 >= 0)
