+++ src/main/java/net/minecraft/world/storage/loot/conditions/RandomChance.java
@@ -32,7 +32,7 @@ public class RandomChance implements LootCondition
 
             public void serialize(JsonObject json, RandomChance value, JsonSerializationContext context)
             {
-                json.addProperty("chance", Float.valueOf(value.chance));
+                json.addProperty("chance", value.chance);
             }
 
             public RandomChance deserialize(JsonObject json, JsonDeserializationContext context)
